+++ components/bootloader_support/src/flash_partitions.c
@@ -28,7 +28,7 @@ static const char *TAG = "flash_parts";
 esp_err_t esp_partition_table_verify(const esp_partition_info_t *partition_table, bool log_errors, int *num_partitions)
 {
     int md5_found = 0;
-    int num_parts;
+    size_t num_parts;
     uint32_t chip_size = g_rom_flashchip.chip_size;
     *num_partitions = 0;
